# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Snap3 Turbo API
  description: |
    Snap3 Turbo briefs → Veo3 preview → QA Gate → exports E2E API
    
    ## Evidence Pack v2 + Audit System
    
    All export endpoints now include comprehensive audit trails and data redaction:
    
    ### Audit Headers
    - `X-Export-SHA256`: SHA256 hash of exported data for integrity verification
    - `X-Export-Size`: Size of payload in bytes
    - `X-Export-Timestamp`: Export timestamp (ISO 8601)
    - `ETag`: Weak ETag for caching based on content hash
    
    ### Data Redaction
    VDP_FULL sensitive data is automatically redacted before export:
    - Removes: overall_analysis, asr_transcript, ocr_text, product_mentions
    - Masks: view_count, like_count, comment_count  
    - Strips: internal/*, debug/*, temp/*, raw/* paths
    
    ### Caching & ETags
    All exports support conditional requests using ETags for bandwidth optimization.
    
    ### Security
    - All exports are logged with audit trail
    - IP addresses are masked for privacy
    - User agents are sanitized 
    - No VDP_FULL data ever exposed in public APIs
  version: 2.0.0
  contact:
    name: Snap3 Team
    email: support@snap3.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.snap3.com
    description: Production server
paths:
...
  /preview/veo:
    post:
      summary: Create preview job
      description: Create async preview job, returns 202 Accepted with Location header
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Unique key for idempotent requests
          schema:
            type: string
        - name: X-Request-Id
          in: header
          required: false
          description: Request ID for tracking and rate limiting
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - veo3Id
                - prompt
                - duration
                - aspectRatio
                - quality
              properties:
                veo3Id:
                  type: string
                  format: uuid
                  description: ID of the Veo3 compilation
                prompt:
                  type: string
                  minLength: 1
                  maxLength: 1000
                  description: Preview generation prompt
                duration:
                  type: number
                  const: 8
                  description: Preview duration (must be 8 seconds)
                aspectRatio:
                  type: string
                  const: '16:9'
                  description: Aspect ratio (must be 16:9)
                quality:
                  type: string
                  enum: [720p, 1080p]
                  description: Video quality
      responses:
        '202':
          description: Preview job accepted
          headers:
            Location:
              description: URL to poll for job status
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                type: object
                required:
                  - jobId
                  - status
                  - pollUrl
                properties:
                  jobId:
                    type: string
                    format: uuid
                    description: Job ID
                  status:
                    type: string
                    enum: [accepted]
                    description: Job status
                  pollUrl:
                    type: string
                    format: uri
                    description: URL to poll for job status
        '400':
          description: Invalid preview request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          description: Too Many Requests
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              schema:
                type: integer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '405':
          description: Method not allowed
  /jobs/{id}:
    get:
      summary: Get job status
      description: Poll job status for completion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Veo3Job'
        '404':
          description: Job not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '405':
          description: Method not allowed
...
components:
  schemas:
...
    Veo3Job:
      type: object
      required:
        - id
        - status
        - progress
        - prompt
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Job ID
        status:
          type: string
          enum: [pending, processing, completed, failed]
          description: Current job status
        progress:
          type: number
          minimum: 0
          maximum: 100
          description: Job progress percentage
        prompt:
          type: string
          description: Original prompt
        attempts:
          type: integer
          minimum: 0
          description: Number of attempts made
        maxAttempts:
          type: integer
          minimum: 1
          description: Maximum number of attempts allowed
        nextAttemptAt:
          type: string
          format: date-time
          description: Next retry attempt timestamp (for failed jobs in backoff)
        createdAt:
          type: string
          format: date-time
          description: Job creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        startedAt:
          type: string
          format: date-time
          description: Processing start timestamp
        completedAt:
          type: string
          format: date-time
          description: Completion timestamp
        result:
          type: object
          description: Job result (for completed jobs)
          properties:
            videoUrl:
              type: string
              format: uri
              description: Generated preview URL
            duration:
              type: number
              const: 8
              description: Video duration
            aspectRatio:
              type: string
              const: '16:9'
              description: Video aspect ratio
            quality:
              type: string
              enum: [720p, 1080p]
              description: Video quality
            fps:
              type: number
              minimum: 24
              maximum: 60
              description: Frames per second
            bitrate:
              type: number
              minimum: 1000000
              description: Video bitrate
            synthIdDetected:
              type: boolean
              description: Whether SynthID was detected
        error:
          type: object
          description: Error details (for failed jobs)
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Error message
            fix:
              type: string
              description: Suggested fix
            retryAfter:
              type: number
              description: Seconds to wait before retry
        history:
          type: array
          description: Job state transition history
          items:
            type: object
            properties:
              from:
                type: string
                enum: [pending, processing, completed, failed]
              to:
                type: string
                enum: [pending, processing, completed, failed]
              timestamp:
                type: string
                format: date-time
        retryAt:
          type: string
          format: date-time
          description: When job can be retried (for failed jobs)
        estimatedCompletionAt:
          type: string
          format: date-time
          description: Estimated completion time (for processing jobs)
...
    Problem:
      type: object
      required:
        - type
        - title
        - status
        - code
      properties:
        type:
          type: string
          format: uri
          description: URI reference that identifies the problem type
        title:
          type: string
          description: Short human-readable summary
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
        instance:
          type: string
          description: URI reference that identifies the specific occurrence
        code:
          type: string
          enum:
            - UNSUPPORTED_AR_FOR_PREVIEW
            - INVALID_DURATION
            - MISSING_FIRST_FRAME
            - PROVIDER_QUOTA_EXCEEDED
            - RATE_LIMITED
            - PROVIDER_POLICY_BLOCKED
            - EMBED_DENIED
            - QA_RULE_VIOLATION
            - VALIDATION_ERROR
            - BAD_REQUEST
            - RESOURCE_NOT_FOUND
            - NOT_FOUND
            - UNAUTHORIZED
            - FORBIDDEN
            - METHOD_NOT_ALLOWED
            - INTERNAL_SERVER_ERROR
            - CONFLICT
            - TIMEOUT
            - SERVICE_UNAVAILABLE
            - INTERNAL_ERROR
          description: Application-specific error code
        fix:
          type: string
          description: Suggested fix for the problem
        retryAfter:
          type: integer
          minimum: 1
          description: Seconds to wait before retrying
        violations:
          type: array
          description: Validation violations
          items:
            type: object
            required:
              - field
              - message
            properties:
              field:
                type: string
                description: Field path that failed validation
              message:
                type: string
                description: Description of the validation failure
              code:
                type: string
                description: Validation error code
        timestamp:
          type: string
          format: date-time
          description: When the problem occurred
        traceId:
          type: string
          description: Unique identifier for request tracing
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - ApiKeyAuth: []